Hints are now usable in Runtime
A sample Hint Creation method has been added in HintManager that debugs the Solved Hints into the OutputLog. Please note DataAssets do not have persistent data- I am working on this now. (Hints are wiped when the Editor is closed) So no Hints will be returned in the GetLevelHints method if you run this without modifying the hint assets first.